Adds TestData class
