+++ src/main/java/Main.java
@@ -1,13 +1,5 @@
 import com.codecool.shop.controller.Controller;
-import com.codecool.shop.dao.ProductCategoryDao;
-import com.codecool.shop.dao.ProductDao;
-import com.codecool.shop.dao.SupplierDao;
-import com.codecool.shop.dao.implementation.ProductCategoryDaoMem;
-import com.codecool.shop.dao.implementation.ProductDaoMem;
-import com.codecool.shop.dao.implementation.SupplierDaoMem;
-import com.codecool.shop.model.Product;
-import com.codecool.shop.model.ProductCategory;
-import com.codecool.shop.model.Supplier;
+import com.codecool.shop.test.*;
 import spark.template.thymeleaf.ThymeleafTemplateEngine;
 
 import static spark.Spark.*;
@@ -21,7 +13,7 @@ public class Main {
         staticFileLocation("/public");
         port(8888);
 
-        populateData();
+        TestData.populateData();
 
         post("/remove/:id", Controller::decreaseItem);
 
@@ -38,35 +30,4 @@ public class Main {
         get("/shoppingcart", Controller::renderCart, new ThymeleafTemplateEngine());
     }
 
-    public static void populateData() {
-
-        ProductDao productDataStore = ProductDaoMem.getInstance();
-        ProductCategoryDao productCategoryDataStore = ProductCategoryDaoMem.getInstance();
-        SupplierDao supplierDataStore = SupplierDaoMem.getInstance();
-
-        //setting up a new supplier
-        Supplier amazon = new Supplier("Amazon", "Digital content and services");
-        supplierDataStore.add(amazon);
-        Supplier lenovo = new Supplier("Lenovo", "Computers");
-        supplierDataStore.add(lenovo);
-        Supplier hp = new Supplier("Hewlett-Packard", "Computers");
-        supplierDataStore.add(hp);
-
-        //setting up a new product category
-        ProductCategory tablet = new ProductCategory("Tablet", "Hardware", "A tablet computer, commonly shortened to tablet, is a thin, flat mobile computer with a touchscreen display.");
-        ProductCategory laptop = new ProductCategory("Laptop", "Hardware", "A laptop computer, flat mobile computer with a display.");
-
-        productCategoryDataStore.add(tablet);
-        productCategoryDataStore.add(laptop);
-
-        //setting up products tablet
-        productDataStore.add(new Product("Amazon Fire", 49.9f, "USD", "Fantastic price. Large content ecosystem. Good parental controls. Helpful technical support.", tablet, amazon));
-        productDataStore.add(new Product("Lenovo IdeaPad Miix 700", 479, "USD", "Keyboard cover is included. Fanless Core m5 processor. Full-size USB ports. Adjustable kickstand.", tablet, lenovo));
-        productDataStore.add(new Product("Amazon Fire HD 8", 89, "USD", "Amazon's latest Fire HD 8 tablet is a great value for media consumption.", tablet, amazon));
-
-        //laptop
-        productDataStore.add(new Product("Amazon Laptop", 111.9f, "USD", "Fantastic price. Large content ecosystem. Good parental controls. Helpful technical support.", laptop, amazon));
-        productDataStore.add(new Product("Lenovo ThinkPad x220", 999, "USD", "Keyboard cover is included. Fanless Core m5 processor. Full-size USB ports. Adjustable kickstand.", laptop, lenovo));
-        productDataStore.add(new Product("HP ProBook", 213, "USD", "Amazon's latest Fire HD 8 tablet is a great value for media consumption.", laptop, hp));
-    }
 }
